+++ src/commands/contract.ts
@@ -10,6 +10,7 @@ import { WasmVmPanel } from '../views/wasmVmPanel';
 export class ContractCmds {
 	public static async Register(context: vscode.ExtensionContext) {
 		this.registerAddContractCmd(context, contractViewProvider);
+		this.registerAddAllContractsCmd(context, contractViewProvider);
 		this.registerSelectContractCmd(context);
 		this.registerDeleteContractCmd(context, contractViewProvider);
 		this.registerUpdateContractAdminCmd(context);
@@ -70,6 +71,56 @@ export class ContractCmds {
 					addr: contractAddr
 				}));
 			}
+		}
+	}
+
+	private static registerAddAllContractsCmd(context: vscode.ExtensionContext, contractViewProvider: ContractDataProvider) {
+		let disposable = vscode.commands.registerCommand('cosmy-wasmy.addContractsAll', () => {
+			if (global.workspaceChain.chainEnvironment == "localnet") {
+				importAllContracts();
+			}
+			else {
+				vscode.window.showErrorMessage(vscode.l10n.t("Sorry! This command is only available for localnet chains. {chain} is marked as '{type}'", {
+					chain: global.workspaceChain.configName,
+					type: global.workspaceChain.chainEnvironment
+				}));
+			}
+		});
+		context.subscriptions.push(disposable);
+
+		function importAllContracts() {
+			vscode.window.withProgress({
+				location: vscode.ProgressLocation.Notification,
+				title: vscode.l10n.t("Fetching all contracts on the chain - {chain}", {
+					chain: global.workspaceChain.configName
+				}),
+				cancellable: false
+			}, (progress, token) => {
+				token.onCancellationRequested(() => { });
+				progress.report({ message: '' });
+				return new Promise((resolve, reject) => {
+					CosmwasmAPI.GetAllContracts().then(resContracts => {
+						let contractsToAdd:Contract[] = [];
+						resContracts.forEach(contract => {
+							if (!Contract.ContractAddressExists(context.globalState, contract.contractAddress)) {
+								contractsToAdd.push(contract);
+							}
+						});
+						Contract.AddManyContract(context.globalState, contractsToAdd);
+						vscode.window.showInformationMessage(vscode.l10n.t("Added {count} contracts!", {
+							count: contractsToAdd.length
+						}));
+						const contracts = Contract.GetContracts(context.globalState);
+						contractViewProvider.refresh(contracts);
+						resolve(contracts);
+					}).catch(err => {
+						vscode.window.showErrorMessage(vscode.l10n.t("Could not import contracts: {err}", {
+							err: err
+						}));
+						reject(err);
+					});
+				});
+			})
 		}
 	}