+++ src/commands/account.ts
@@ -33,7 +33,7 @@ export class AccountCmds {
 						vscode.window.showQuickPick(options).then(rr => {
 							if (rr) {
 								if (rr === vscode.l10n.t("Generate seed phrase for me (Recommended)")) {
-									let defaultLen = WrapWallet.isEthSecp256k1(global.workspaceChain.signType) ? 12 : 24;
+									let defaultLen = global.workspaceChain.signType === Constants.SIGN_TYPE.ethsecp256k1 ? 12 : 24;
 									WrapWallet.generate(global.workspaceChain.signType, defaultLen).then(wallet => {
 										const account = new Account(accountLabel, wallet.mnemonic);
 										saveNewAccount(account);