+++ CHANGELOG.md
@@ -18,6 +18,25 @@ All notable changes to the Cosmy Wasmy extension will be documented in this file
 ### Security 
 -->
 
+## [Unreleased]
+
+### Added
+
+- For localnet chains, all contracts on the chain can be imported in single click.
+- Support [Finschia testnet](https://www.finschia.io/) - Finschia aims to become the world’s No. 1 blockchain ecosystem, leading the popularization of Web3 and securing over 1 billion users. Added configurations for localnet as well as testnet chains
+- Support `coinType` - Can use it when generate account, default `118`.
+
+### Changed
+
+### Deprecated
+
+### Removed
+
+### Fixed
+
+### Security 
+
+
 ## [v2.3.1] - 07 July 2023
 
 ### Removed