+++ CHANGELOG.md
@@ -34,7 +34,7 @@ All notable changes to the Cosmy Wasmy extension will be documented in this file
 
 ### Fixed
 
-- Fix default configuration
+- Fix default chain name
 
 ### Security