+++ CHANGELOG.md
@@ -34,6 +34,7 @@ All notable changes to the Cosmy Wasmy extension will be documented in this file
 
 ### Fixed
 
+- Fix default chain name
 - Fix auto complete for latest cosmwasm version
 
 ### Security