+++ CHANGELOG.md
@@ -34,7 +34,9 @@ All notable changes to the Cosmy Wasmy extension will be documented in this file
 
 ### Fixed
 
-- refactor workspace and chainConfig class
+- Fix default chain name
+- Fix auto complete for latest cosmwasm version
+- Refactor workspace and chainConfig class
 
 ### Security